refactor: un-Class-ify the `Class` model & separate scraper's logic out!
