+++ common/src/model/Class.ts
@@ -1,9 +1,6 @@
-/* eslint-disable @typescript-eslint/no-use-before-define */
-/* eslint-disable max-classes-per-file */
-/* eslint-disable lines-between-class-members */
+/* eslint-disable import/no-cycle */
 
-import { getSpecificScheduleURI } from "./Schedule";
-// eslint-disable-next-line import/no-cycle
+import { Participant } from "./Participant";
 import { Lesson } from "./Lesson";
 
 export const classNumMin = 0;
@@ -11,11 +8,6 @@ export const classNumMax = 12;
 
 /** Use `0` ONLY FOR FAILED CASES */
 export type TClassNum = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
-
-interface ClassInitData extends Partial<Class> {
-	text: string;
-	originalHref: string;
-}
 
 /**
  * A `class` describes all students within a particular class / grade,
@@ -25,9 +17,9 @@ interface ClassInitData extends Partial<Class> {
  * TODO: https://github.com/sarpik/turbo-schedule/issues/36
  *
  */
-export interface Class {
+export interface Class extends Participant {
 	// 	/** BEGIN Scrapable */
-	// readonly id: string; /** TODO */
+	readonly id: string /** TODO */;
 	readonly text: string;
 	readonly originalHref: string;
 	readonly originalScheduleURI: string;
@@ -46,134 +38,20 @@ export interface Class {
 	/** END Class */
 }
 
-export function createClass(data: ClassInitData = { text: "", originalHref: "" }): Class {
-	const { text, originalHref } = data;
+export const getDefaultClass = (): Class => ({
+	id: "",
+	labels: [],
+	text: "",
+	originalHref: "",
+	originalScheduleURI: "",
 
-	const fullClassOrig = text;
-	const originalScheduleURI = getSpecificScheduleURI(originalHref);
+	fullClassOrig: "",
+	classNumOrig: "",
+	classCharOrig: "",
 
-	const classNumOrig = parseClassNumOrig(fullClassOrig);
-	const classCharOrig = parseClassCharOrig(fullClassOrig);
+	fullClass: "",
+	classNum: 0,
+	classChar: "",
 
-	const fullClass = parseFullClass(fullClassOrig);
-	const classNum = parseClassNum(fullClassOrig);
-	const classChar = parseClassChar(fullClassOrig);
-
-	const theClass: Class = {
-		text,
-		originalHref,
-		originalScheduleURI,
-		//
-		fullClassOrig,
-		classNumOrig,
-		classCharOrig,
-		//
-		fullClass,
-		classNum,
-		classChar,
-	};
-
-	return theClass;
-}
-
-// class Class {
-// 	/** BEGIN Scrapable */
-// 	// readonly id: string; /** TODO */
-
-// 	readonly text: string;
-// 	readonly originalHref: string;
-
-// 	readonly originalScheduleURI: string;
-// 	/** END Scrapable */
-
-// 	/** BEGIN Class */
-// 	readonly classNum: TClassNum;
-// 	readonly classChar: string;
-// 	readonly fullClass: string;
-
-// 	readonly classNumOrig: string;
-// 	readonly classCharOrig: string;
-// 	readonly fullClassOrig: string;
-// 	/** END Class */
-
-// 	constructor(data?: ClassInitData) {
-// 		// const class = {
-// 		// 	text,
-// 		// 	originalHref,
-// 		// 	originalScheduleURI: getSpecificScheduleURI(originalHref),
-// 		// 	// classNum:
-// 		// };
-// 	}
-
-// }
-
-function parseClassCharOrig(fullClassOrig: string): string {
-	return fullClassOrig[fullClassOrig.length - 1];
-}
-
-function parseClassNumOrig(fullClassOrig: string): string {
-	return fullClassOrig.slice(0, -1);
-}
-
-//
-function parseClassChar(fullClassOrig: string): string {
-	return parseClassCharOrig(fullClassOrig);
-}
-
-const retardedClassNumToNormalClassNumDict = {
-	i: 9,
-	ig: 9,
-	ii: 10,
-	iig: 10,
-	iii: 11,
-	iiig: 11,
-	iv: 12,
-	ivg: 12,
-};
-
-type TRetardedClassNum = typeof retardedClassNumToNormalClassNumDict;
-
-function parseRetardedClassNum(fullClassNameOrig: string): number {
-	return retardedClassNumToNormalClassNumDict[fullClassNameOrig.toLowerCase() as keyof TRetardedClassNum];
-}
-
-function parseClassNum(fullClassOrig: string): TClassNum {
-	console.log("parseClassNum: fullClassOrig:", fullClassOrig);
-
-	const classNumStr: string = fullClassOrig /**  "6a" or "IIGc" */
-		.slice(0, -1); /**                         "6"  or "IIG"  */
-
-	const classNumDangerous: number = classNumStr.toLowerCase().includes("i")
-		? parseRetardedClassNum(classNumStr)
-		: Number(classNumStr);
-
-	if (!Number.isInteger(classNumDangerous)) {
-		throw new Error(
-			`\
-\`classNum\` is NOT an integer!
-
-provided = \`${fullClassOrig}\`
-parsed & invalid = \`${classNumDangerous}\`,
-typeof = \`${typeof classNumDangerous}\``
-		);
-	}
-
-	if (classNumDangerous < classNumMin || classNumDangerous > classNumMax) {
-		throw new Error(
-			`\
-\`classNum\` was out of range!
-
-min = \`${classNumMin}\`,
-max = \`${classNumMax}\`,
-provided = \`${classNumDangerous}\``
-		);
-	}
-
-	const classNumSafe: TClassNum = classNumDangerous as TClassNum;
-
-	return classNumSafe;
-}
-
-function parseFullClass(fullClassOrig: string): string {
-	return parseClassNum(fullClassOrig) + parseClassChar(fullClassOrig);
-}
+	lessons: [],
+});
